Switch to using jekyll-assets to include assets
This commit uses the jekyll-assets gem to generate nice asset paths
that are cachebusted properly.

Rename all the JavaScript and Stylesheets to put them in the _assets
directory. Put the toolkit submodule there too.

Remove stray image-url() in _typography.scss and _buttons.scss

Use the provided liquid tags to include the CSS and JS in the HTML
document.