Update govuk_template and footer links
The goal of this was to update the footer links. However, the footer
styling had been updated in govuk_template so trying to copy the styles
from static for the new links caused layout problems. So I have also
updated govuk_template to the latest version. There were a couple of
small style changes and new apple touch icons that came with the update.